feat: Add Farm feature with TDD level system
  - Add FarmScreen with foraging, fishing, mining activities
  - Refactor level to computed property (TDD: single source of truth)
  - Add ScreenTopBar reusable component
  - Add ElfInfoTests for level calculations
  - Change Int16 to Int for level types

+++ Packages/elf_Kit/Sources/DataLayer/Builders/CombatantSnapshotBuilder/CombatantSnapshotBuilder.swift
@@ -15,7 +15,7 @@ public protocol CombatantSnapshotBuilder: Sendable {
     /// - Parameters:
     ///   - name: Display name for the combatant
     ///   - imageName: Image asset name for UI display
-    ///   - level: Hero level
+    ///   - level: Hero level (1-12)
     ///   - fightStyleAttributes: Attributes from selected fight style
     ///   - randomLevelAttributes: Random attributes gained from levels
     ///   - selectedItems: Dictionary of selected item UUIDs by type
@@ -23,7 +23,7 @@ public protocol CombatantSnapshotBuilder: Sendable {
     func buildSnapshot(
         name: String,
         imageName: String,
-        level: Int16,
+        level: Int,
         fightStyleAttributes: HeroAttributes,
         randomLevelAttributes: HeroAttributes,
         selectedItems: [HeroItemType: UUID?]
