+++ Packages/elf_Kit/Sources/DataLayer/Services/House/HouseService.swift
@@ -16,20 +16,20 @@ public protocol HouseService: Sendable {
     /// Create a house with AI elves only
     /// - Parameters:
     ///   - templateIndex: Index of the house template (0-7)
-    ///   - level: Level for AI elves
+    ///   - level: Level for AI elves (1-12)
     /// - Returns: House with 10 AI elves
-    func createHouse(templateIndex: Int, level: Int16) async -> House
+    func createHouse(templateIndex: Int, level: Int) async -> House
 
     /// Create a house with player and AI elves
     /// - Parameters:
     ///   - templateIndex: Index of the house template (0-7)
-    ///   - level: Level for AI elves
+    ///   - level: Level for AI elves (1-12)
     ///   - playerElfInfo: Player's elf info
     ///   - playerMemberIndex: Position for player in the house (0-9)
     /// - Returns: House with player and 9 AI elves
     func createHouse(
         templateIndex: Int,
-        level: Int16,
+        level: Int,
         playerElfInfo: ElfInfo,
         playerMemberIndex: Int
     ) async -> House
